Atualizações conforme solicitações na pull request
