+++ frontend/src/utils/getFirstName.ts
@@ -1,0 +1,4 @@
+export function getFirstName(name: string): string {
+  const names = name.split(" ");
+  return names[0];
+}
